[refactor] 악보 연습 채점 기능 분리 및 dto lombok 설정 수정 - AudioMessageConsumer에서 채점 관련 기능을 PracticeResultResolver로 분리 - dto에 builder 추가 및 접근 권한 수정
